Changes to practice priority

+++ src/MyArrays.java
@@ -1,12 +1,12 @@
 // quickSort.java sorts an array in decreasing order
 
 public class MyArrays {
-   public static int[] shellSort(float[] a, int fromIndex, int toIndex) {
+   public static int[] shellSort(double[] a, int fromIndex, int toIndex) {
 
        int[] Index = new int[toIndex];
        int d;
 
-      int h = 1; 
+      int h = 1;
 
       for (int k=0; k<toIndex; k++)
 	  Index[k] = k;
@@ -17,7 +17,7 @@ public class MyArrays {
          h = h/3;
          for (int i=fromIndex+h; i<toIndex; i++) {
             d = Index[fromIndex+i];
-            int j = i; 
+            int j = i;
             while (j>=fromIndex+h && a[d] < a[Index[fromIndex+j-h]]) {
                Index[fromIndex+j] = Index[fromIndex+j-h];
                j = j - h;
@@ -29,12 +29,12 @@ public class MyArrays {
    }
 
 
-   public static int[] shellSortDec(float[] a, int fromIndex, int toIndex) {
+   public static int[] shellSortDec(double[] a, int fromIndex, int toIndex) {
 
        int[] Index = new int[toIndex];
        int d;
 
-      int h = 1; 
+      int h = 1;
 
       for (int k=0; k<toIndex; k++)
 	  Index[k] = k;
@@ -45,7 +45,7 @@ public class MyArrays {
          h = h/3;
          for (int i=fromIndex+h; i<toIndex; i++) {
             d = Index[fromIndex+i];
-            int j = i; 
+            int j = i;
             while (j>=fromIndex+h && a[d] >= a[Index[fromIndex+j-h]]) {
                Index[fromIndex+j] = Index[fromIndex+j-h];
                j = j - h;
